LayoutTests/storage/indexeddb/mozilla/bad-keypath-expected.txt

Test IndexedDB adding property with invalid keypath

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('foo', { keyPath: 'keyPath' });
request = objectStore.add({ keyPath: 'foo' });
Expecting exception from request = objectStore.add({});
PASS Exception was thrown.
PASS code is IDBDatabaseException.DATA_ERR
PASS successfullyParsed is true

TEST COMPLETE
